third_party/blink/web_tests/external/wpt/css/css-values/urls/url-request-modifiers-serialize.sub-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" cross-origin(anonymous))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" cross-origin(use-credentials))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" integrity(\\"sha384-foobar\\"))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" integrity(\\"\\"))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" referrer-policy(no-referrer))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" referrer-policy(no-referrer-when-downgrade))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" referrer-policy(same-origin))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" referrer-policy(origin))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" referrer-policy(strict-origin))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" referrer-policy(origin-when-cross-origin))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" referrer-policy(strict-origin-when-cross-origin))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" referrer-policy(unsafe-url))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" cross-origin(anonymous) integrity(\\"sha384-foobar\\"))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" integrity(\\"sha384-foobar\\") cross-origin(anonymous))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" integrity(\\"sha384-foobar\\") referrer-policy(no-referrer))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" referrer-policy(no-referrer) integrity(\\"sha384-foobar\\"))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" cross-origin(anonymous) referrer-policy(no-referrer))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" referrer-policy(no-referrer) cross-origin(anonymous))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" cross-origin(anonymous) integrity(\\"sha384-foobar\\") referrer-policy(no-referrer))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" integrity(\\"sha384-foobar\\") referrer-policy(no-referrer) cross-origin(anonymous))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" referrer-policy(no-referrer) cross-origin(anonymous) integrity(\\"sha384-foobar\\"))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-image'] = "url(\\"http://web-platform.test:8001/css/support/1x1-green.png\\" referrer-policy(no-referrer) integrity(\\"sha384-foobar\\") cross-origin(anonymous))" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
